prototype cleanup bsr state
