+++ examples/plate/1_linear_static.cu
@@ -26,15 +26,34 @@ int main() {
     using ElemGroup = ShellElementGroup<T, Director, Basis, Physics>;
     using Assembler = ElementAssembler<T, ElemGroup, VecType, BsrMat>;
 
-    int nxe = 300; // 100
+    int nxe = 3;
+    // int nxe = 100;
+    // int nxe = 300;
     int nye = nxe;
     double Lx = 2.0, Ly = 1.0, E = 70e9, nu = 0.3, thick = 0.005;
     auto assembler = createPlateAssembler<Assembler>(nxe, nye, Lx, Ly, E, nu, thick);
 
-    // BSR factorization
-    double fillin = 10.0; // 10.0
+    // BSR symbolic factorization
+    auto bsr_data = assembler.getBsrData();
+    double fillin = 10.0;  // 10.0
     bool print = true;
-    assembler.symbolic_factorization(fillin, print);
+    bsr_data.compute_full_LU_pattern(fillin, print);
+
+    printf("rowp:");
+    printVec<int>(bsr_data.nnodes+1, bsr_data.rowp);
+    printf("cols:");
+    printVec<int>(bsr_data.nnzb, bsr_data.cols);
+    printf("perm:");
+    printVec<int>(bsr_data.nnodes, bsr_data.perm);
+    printf("iperm:");
+    printVec<int>(bsr_data.nnodes, bsr_data.iperm);
+    printf("elem_ind_map:");
+    printVec<int>(bsr_data.nelems * bsr_data.nodes_per_elem * bsr_data.nodes_per_elem, 
+        bsr_data.elem_ind_map);
+
+    assembler.moveBsrDataToDevice();
+
+    printf("here1\n");
 
     // get the loads
     double Q = 1.0; // load magnitude
@@ -43,19 +62,28 @@ int main() {
     auto loads = assembler.createVarsVec(my_loads);
     assembler.apply_bcs(loads);
 
+    printf("here2\n");
+
     // setup kmat and initial vecs
     auto kmat = createBsrMat<Assembler, VecType<T>>(assembler);
     auto soln = assembler.createVarsVec();
     auto res = assembler.createVarsVec();
     auto vars = assembler.createVarsVec();
 
+    printf("here3\n");
+
     // assemble the kmat
     assembler.add_jacobian(res, kmat);
+    printf("here3.5\n");
     assembler.apply_bcs(res);
+    printf("here3.6\n");
     assembler.apply_bcs(kmat);
+    printf("here4\n");
 
     // solve the linear system
     CUSPARSE::direct_LU_solve(kmat, loads, soln);
+
+    printf("here5\n");
 
     // print some of the data of host residual
     auto h_soln = soln.createHostVec();
