+++ examples/uCRM/1_linear_static.cu
@@ -39,9 +39,11 @@ int main() {
   auto assembler = Assembler::createFromBDF(mesh_loader, Data(E, nu, thick));
 
   // BSR factorization
+  auto bsr_data = assembler.getBsrData();
   double fillin = 10.0;  // 10.0
   bool print = true;
-  assembler.symbolic_factorization(fillin, print);
+  bsr_data.compute_full_LU_pattern(fillin, print);
+  assembler.moveBsrDataToDevice();
 
   // get the loads
   int nvars = assembler.get_num_vars();
