+++ include/linalg/_linalg.h
@@ -1,6 +1,5 @@
 #pragma once
 
 #include "bsr_mat.h"
-#include "dense_mat.h"
 #include "precond.h"
 #include "vec.h"
